requirements.txt
SQLAlchemy==1.3.3
sqlalchemy_utils==0.33.9
pymysql==0.9.3
python-dotenv~=0.10.1
elasticsearch
Flask
requests